Move 'componentWillMount' to constructor
'componentWillMount' is replaced by an anonymous self-invoking function in the constructor.
We use a new function because 'this.initGrid' itself cannot be called in the constructor,
since it uses 'setState'. Neither can 'this.initGrid' be moved into 'componentDidMount',
because mounting the component requires an initialised grid.
To avoid confusion, we rename 'this.initGrid' to 'this.resetGrid', since the actual
initialisation now happens in the constructor. The renamed function, 'this.resetGrid',
is now only used to modify the grid through the Gridsize settings component.