Remove anonymous initialisation function in constructor
This function adds a whole lot of code for doing something relatively simple,
which can just as well be hardcoded. Hardcoding in this case helps in visualising
the arrays, is more efficient, and removes code bloat. We do lose out on customisability,
but these state variables are not intended to be manually customised anyway ...